Category + product group
Fix errors

+++ src/main/java/com/localbrand/dal/repository/ICategoryRepository.java
@@ -9,5 +9,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 @Repository
 public interface ICategoryRepository extends JpaRepository<Category, String>{
 	
-	long countByName(String name);
+	int countByName(String name);
 }
